Update wallsolverCPU.cu
Added the visual display function. Function is named: displayBoard() and it is currently called in main right under the outputBoard() function.

Takes a nxn array and scales it to a 1+2nx2n array in order to fit the visual display medium, but other than having to mentally add 2 to each space ([1][1] becomes [3][3]) it should be straightforward. Let me know if you have questions.